updated examinations section
